Rename to module.rb
